+++ src/main/java/com/bisnode/opa/StopOpaTask.java
@@ -14,7 +14,7 @@ public class StopOpaTask extends DefaultTask {
     public void stopOpa() {
         boolean result = OpaPluginUtils.stopOpaProcess(getProject());
         if (getLogger().isDebugEnabled()) {
-            getLogger().debug(result ? "OPA stopped" : "Did not find OPA process to stop");
+            getLogger().debug(result ? "OPA stopped" : "Did not find OPA process to stop or the process is hanging");
         }
     }