+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [//]: # (### Changed)
 [//]: # (### Removed)
 
+## [0.3.1] - 2020-11-23
+### Changed
+- GH-28 OPA process' output is now consumed and logged on info level, so the subprocess doesn't hang anymore
+- startOpa task now waits for OPA server to initialize
+- enabled test in `StopOpaTaskTest`
+- for now, tests rely on `opa` in `$PATH`
+
 ## [0.3.0] - 2020-07-16
 ### Added
 - GH-15 Make testRego task output `opa -v`-like summary when invoked with `--info`